+++ src/components/index.ts
@@ -2,11 +2,11 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './Alert/index';
 export * from './Button/index';
 export * from './Empty/index';
 export * from './GlobalNavigation/index';
 export * from './Layout/index';
+export * from './Overlay/index';
 export * from './Product/index';
 export * from './ProductDetail/index';
 export * from './Typography/index';
